Refactor tests to share Home Assistant stubs
