+++ tests/test_update_listener.py
@@ -1,23 +1,16 @@
 import os
 import sys
 import importlib
-import types
 from types import SimpleNamespace
 
 import pytest
 
+sys.path.insert(0, os.path.dirname(__file__))
+from hass_stubs import install_homeassistant_stubs
+
+install_homeassistant_stubs()
+
 BASE_DIR = os.path.dirname(os.path.dirname(__file__))
-
-# Stub minimal Home Assistant modules
-ha = types.ModuleType("homeassistant")
-ha.config_entries = types.ModuleType("config_entries")
-ha.core = types.ModuleType("core")
-ha.config_entries.ConfigEntry = object
-ha.core.HomeAssistant = object
-sys.modules.setdefault("homeassistant", ha)
-sys.modules.setdefault("homeassistant.config_entries", ha.config_entries)
-sys.modules.setdefault("homeassistant.core", ha.core)
-
 sys.path.insert(0, BASE_DIR)
 init = importlib.import_module("custom_components.openai_gpt4o_tts.__init__")
